rename pkg, add tests, and use core logger
